Configures services for staging environment
Updates Docker Compose configuration to use a staging environment.

This change configures the API and workout generator services to
utilize the staging environment variables by mounting .env files as volumes,
setting the ASPNETCORE_ENVIRONMENT to Staging, and including a healthcheck for RabbitMQ.

The RabbitMQ password has been updated to a default password, and the
dotenv-vault files have also been updated.
A README has been added with instructions to get started.